moved create_ai_function_info to function_context for reusability and reduce repetation
